Fixup issue with dependencies
Seems like something went terribly wrong and the author of
gomodules/jsonpatch moved a tag from one commit to another. This broker
everything (see info in:
https://github.com/gomodules/jsonpatch/issues/21).

As a result, and because of the advice in this issue:
https://github.com/kubernetes-sigs/kubebuilder/issues/997 we have bumped
our controller-runtime dependency to resolve the issue.

This seemed to work. Also bumped the go version in go.mod.

Co-authored-by: James Myers <jmyers@pivotal.io>