Set CI tasks to use 'latest' CI image tag
We had a discussion in tech forum and decided to switch to using the
'latest' tag, rather than a versioned tag. The main reason was to reduce
management overhead and to speed up bumping the image.

We spun out a separate story (#169149008) in order to ensure we have
traceability Re: "What version of package X  was used to build
Marketplace version Y?".

[#169149003]