Add basic support for multiple role rules
NB: This is a super basic approach to adding support for multiple role
rules. With this commit, the projects-operator will now create 1 role
with 3 distinct rules (it will fail if there aren't exactly 3!).

Context as to why we have taken this approach and the long-term plan for
adding "proper" support for this are in the story.

Ps: This commit is not intended to land on master, but to live on a
short-term branch, which we will use for the marketplace alpha.

[#169185275]

Co-authored-by: Glen Rodgers <grodgers@pivotal.io>